rusk: rename `prover` module in `verifier`
