+++ rusk/src/lib/vm.rs
@@ -124,7 +124,7 @@ impl VMExecution for Rusk {
             let err = crate::Error::RepeatingNullifiers(existing_nullifiers);
             return Err(anyhow::anyhow!("Invalid tx: {err}"));
         }
-        match crate::prover::verify_proof(tx) {
+        match crate::verifier::verify_proof(tx) {
             Ok(true) => Ok(()),
             Ok(false) => Err(anyhow::anyhow!("Invalid proof")),
             Err(e) => Err(anyhow::anyhow!("Cannot verify the proof: {e}")),
